Functions with API at 100%, with DB at 90%
