+++ api/src/controllers/backupController.js
@@ -1,24 +1,5 @@
-const axios = require ("axios")
 
-async function getApiData (req, res){
-    try {
-        let dataDefault = await axios(`https://api.itbook.store/1.0/search/mongo`)
-        let apiData = await dataDefault.data.books.map(b => {
-            return {
-                title: b.title,
-                subtitle: b.subtitle,
-                isbn13: b.isbn13,
-                price: b.price,
-                image: b.image
-            }
-        })
-        return apiData
-    } catch (error) {
-        console.log({error: "err"})
-    }
-}
-
-
+//SAVE BOOKS IN DB
 async function getBooksByName (req, res){
     const { name } = req.params
     try {
@@ -26,7 +7,6 @@ async function getBooksByName (req, res){
         for (let i = 0; i < 10; i++) {
             let data = await axios(`https://api.itbook.store/1.0/search/${name}/${i}`)
             bookSearched.push(await data.data.books.map(b => b))
-            
         }
         let dataDefault = bookSearched.flat()
         let apiData = dataDefault.map(b => {
@@ -38,7 +18,6 @@ async function getBooksByName (req, res){
                 image: b.image
             }
         })
-        console.log(apiData)
         Book.insertMany(apiData, (errors, insertedBooks) => {
             if(errors) {
               return res.status(400).json({ok: false, errors});
@@ -52,54 +31,7 @@ async function getBooksByName (req, res){
 }
 
 
-
-
-
-
-
-
-
-
-
-
-
-async function getBooksByName (req, res){
-    const { name } = req.params
-    try {
-        let bookSearched = []
-        for (let i = 0; i < 10; i++) {
-            let data = await axios(`https://api.itbook.store/1.0/search/${name}/${i}`)
-            bookSearched.push(await data.data.books.map(b => b))
-            
-        }
-        let dataDefault = bookSearched.flat()
-        let apiData = dataDefault.map(b => {
-            return {
-                title: b.title,
-                subtitle: b.subtitle,
-                isbn13: b.isbn13,
-                price: b.price,
-                image: b.image
-            }
-        })
-        // console.log(apiData)
-        Book.insertMany(apiData, (errors, insertedBooks) => {
-            if(errors) {
-              return res.status(400).json({ok: false, errors});
-            }
-        })
-
-        // const bookSaved = await booksOnDb.save()
-                res.status(201).send("ok") 
-    } catch (error) {
-        res.status(404).json({error: "An unexpected error occurred, please try again later"})
-    }
-}
-
-
-
-
-
+//TRY TO SAVE MORE DETAILS INTO DB
 async function getBooksById (req, res){
     // const { id } = req.params
     try {
@@ -131,9 +63,6 @@ async function getBooksById (req, res){
             return null
             console.log(arrayNoexiste)
         }) 
-        
-        
-        // let book = await axios(`https://api.itbook.store/1.0/books/${id}`)
         res.status(200).send(array) 
     } catch (error) {
         res.status(404).json({error: "An unexpected error occurred, please try again later"})
@@ -141,25 +70,24 @@ async function getBooksById (req, res){
 }
 
 
-
-// let todo = await Book.find({title: /jira/i})
-    // const tada = todo.map(b => b.isbn13)
-
-    
 //REMOVER DUPLICATEDDDDD
-    // const todo = await Book.aggregate([
-    //     { $group: {
-    //         _id: { isbn13: "$isbn13" },
-    //         dups: { "$push": "$_id" },
-    //         count: { "$sum": 1 }
-    //     }},
-    //     { $match: { "count": { "$gt": 1 } }}
-    // ])
-    // todo.forEach(async (doc) => {
-    //     doc.dups.shift();
-    //     await Book.remove({ "_id": {"$in": doc.dups }});
-    // });
-    // console.log(todo);
+// let allBooks = await Book.find({}).select("isbn13")
+// let todo = await Book.find({title: /jira/i})
+//     const tada = todo.map(b => b.isbn13)
+    
+//     const todo = await Book.aggregate([
+//         { $group: {
+//             _id: { isbn13: "$isbn13" },
+//             dups: { "$push": "$_id" },
+//             count: { "$sum": 1 }
+//         }},
+//         { $match: { "count": { "$gt": 1 } }}
+//     ])
+//     todo.forEach(async (doc) => {
+//         doc.dups.shift();
+//         await Book.remove({ "_id": {"$in": doc.dups }});
+//     });
+//     console.log(todo);
